fix: XF tts model
