+++ apps/models_provider/impl/xf_model_provider/credential/tts/__init__.py
@@ -1,0 +1,11 @@
+# coding=utf-8
+"""
+    @project: MaxKB
+    @Author：niu
+    @file： __init__.py.py
+    @date：2025/12/10 14:13
+    @desc:
+"""
+from .tts import *
+from .default_tts import *
+from .super_humanoid_tts import *
